+++ UIComponents.Web.Tests/UIComponents/CHANGELOG.md
@@ -1,4 +1,7 @@
-﻿# 2024-06-06 (1.2.10)
+﻿# 2024-06-13 (1.2.13)
+- JsGrid Control can filter edit or delete button
+- Fix selectlist items for nullable enum﻿# 2024-06-10 (1.2.11)
+- Select list renderer voor UICInputTime﻿# 2024-06-06 (1.2.10)
 - Fix filter compabiliteit UICTable met oude JsGrid taghelper van CdcPortal﻿# 2024-05-31 (1.2.09)
 - UICTable rowrenderer fix
 - UICActionRefresh fix